move ip utils to their own file
