+++ app/util/str.spec.ts
@@ -5,9 +5,9 @@
  *
  * Copyright Oxide Computer Company
  */
-import { describe, expect, it, test } from 'vitest'
+import { describe, expect, it } from 'vitest'
 
-import { camelCase, capitalize, commaSeries, kebabCase, titleCase, validateIp } from './str'
+import { camelCase, capitalize, commaSeries, kebabCase, titleCase } from './str'
 
 describe('capitalize', () => {
   it('capitalizes the first letter', () => {
@@ -75,68 +75,4 @@ describe('titleCase', () => {
   it('doesn’t modify non-letter characters', () => {
     expect(titleCase('123 abc')).toBe('123 Abc')
   })
-})
-
-// Rust playground comparing results with std::net::{Ipv4Addr, Ipv6Addr}
-// https://play.rust-lang.org/?version=stable&mode=debug&edition=2021&gist=493b3345b9f6c0b1c8ee91834e99ef7b
-
-test.each(['123.4.56.7', '1.2.3.4'])(
-  'validateIp catches valid IPV4 / invalid IPV6: %s',
-  (s) => {
-    expect(validateIp(s)).toStrictEqual({ isv4: true, isv6: false, valid: true })
-  }
-)
-
-test.each([
-  '2001:db8:3333:4444:5555:6666:7777:8888',
-  '2001:db8:3333:4444:CCCC:DDDD:EEEE:FFFF',
-  '::',
-  '2001:db8::',
-  '::1234:5678',
-  '2001:db8::1234:5678',
-  '2001:0db8:85a3:0000:0000:8a2e:0370:7334',
-  '::ffff:192.0.2.128',
-  '1:2:3:4:5:6:7:8',
-  '::ffff:10.0.0.1',
-  '::ffff:1.2.3.4',
-  '::ffff:0.0.0.0',
-  '1:2:3:4:5:6:77:88',
-  '::ffff:255.255.255.255',
-  'fe08::7:8',
-  'ffff:ffff:ffff:ffff:ffff:ffff:ffff:ffff',
-])('validateIp catches invalid IPV4 / valid IPV6: %s', (s) => {
-  expect(validateIp(s)).toStrictEqual({ isv4: false, isv6: true, valid: true })
-})
-
-test.each([
-  '',
-  '1',
-  'abc',
-  'a.b.c.d',
-  // some implementations (I think incorrectly) allow leading zeros but nexus does not
-  '01.102.103.104',
-  '127.0.0',
-  '127.0.0.1.',
-  '127.0.0.1 ',
-  ' 127.0.0.1',
-  '10002.3.4',
-  '1.2.3.4.5',
-  '256.0.0.0',
-  '260.0.0.0',
-  '256.1.1.1',
-  '2001:0db8:85a3:0000:0000:8a2e:0370:7334 ',
-  ' 2001:db8::',
-  '1:2:3:4:5:6:7:8:9',
-  '1:2:3:4:5:6::7:8',
-  ':1:2:3:4:5:6:7:8',
-  '1:2:3:4:5:6:7:8:',
-  '::1:2:3:4:5:6:7:8',
-  '1:2:3:4:5:6:7:8::',
-  '1:2:3:4:5:6:7:88888',
-  '2001:db8:3:4:5::192.0.2.33', // std::new::Ipv6Net allows this one
-  'fe08::7:8%',
-  'fe08::7:8i',
-  'fe08::7:8interface',
-])('validateIp catches invalid IP: %s', (s) => {
-  expect(validateIp(s)).toStrictEqual({ isv4: false, isv6: false, valid: false })
 })
